requirements.txt
numpy
py-cpuinfo

# optional
# yattag